+++ src/python/index.ts
@@ -1,6 +1,7 @@
 export * as bisect from './bisect/index.ts'
 export * as calendar from './calendar/index.ts'
 export * as difflib from './difflib/index.ts'
+export * as functools from './functools/index.ts'
 export * as itertools from './itertools/index.ts'
 export * as math from './math/index.ts'
 export * as operator from './operator/index.ts'